+++ tests/Unit/Blob/Models/CopyBlobOptionsTest.php
@@ -41,6 +41,8 @@ use MicrosoftAzure\Storage\Blob\Models\CopyBlobOptions;
 class CopyBlobOptionsTest extends \PHPUnit_Framework_TestCase
 {
     /**
+     * @covers MicrosoftAzure\Storage\Blob\Models\CopyBlobFromURLOptions::setMetadata
+     * @covers MicrosoftAzure\Storage\Blob\Models\CopyBlobFromURLOptions::getMetadata
      * @covers MicrosoftAzure\Storage\Blob\Models\CopyBlobOptions::setMetadata
      * @covers MicrosoftAzure\Storage\Blob\Models\CopyBlobOptions::getMetadata
      */
@@ -57,6 +59,8 @@ class CopyBlobOptionsTest extends \PHPUnit_Framework_TestCase
     }
     
     /**
+     * @covers MicrosoftAzure\Storage\Blob\Models\CopyBlobFromURLOptions::setAccessConditions
+     * @covers MicrosoftAzure\Storage\Blob\Models\CopyBlobFromURLOptions::getAccessConditions
      * @covers MicrosoftAzure\Storage\Blob\Models\CopyBlobOptions::setAccessConditions
      * @covers MicrosoftAzure\Storage\Blob\Models\CopyBlobOptions::getAccessConditions
      */
@@ -73,6 +77,8 @@ class CopyBlobOptionsTest extends \PHPUnit_Framework_TestCase
     }
     
     /**
+     * @covers MicrosoftAzure\Storage\Blob\Models\CopyBlobFromURLOptions::setSourceAccessConditions
+     * @covers MicrosoftAzure\Storage\Blob\Models\CopyBlobFromURLOptions::getSourceAccessConditions
      * @covers MicrosoftAzure\Storage\Blob\Models\CopyBlobOptions::setSourceAccessConditions
      * @covers MicrosoftAzure\Storage\Blob\Models\CopyBlobOptions::getSourceAccessConditions
      */
@@ -89,6 +95,8 @@ class CopyBlobOptionsTest extends \PHPUnit_Framework_TestCase
     }
     
     /**
+     * @covers MicrosoftAzure\Storage\Blob\Models\CopyBlobFromURLOptions::setLeaseId
+     * @covers MicrosoftAzure\Storage\Blob\Models\CopyBlobFromURLOptions::getLeaseId
      * @covers MicrosoftAzure\Storage\Blob\Models\CopyBlobOptions::setLeaseId
      * @covers MicrosoftAzure\Storage\Blob\Models\CopyBlobOptions::getLeaseId
      */
@@ -102,6 +110,8 @@ class CopyBlobOptionsTest extends \PHPUnit_Framework_TestCase
     }
     
     /**
+     * @covers MicrosoftAzure\Storage\Blob\Models\CopyBlobFromURLOptions::setSourceLeaseId
+     * @covers MicrosoftAzure\Storage\Blob\Models\CopyBlobFromURLOptions::getSourceLeaseId
      * @covers MicrosoftAzure\Storage\Blob\Models\CopyBlobOptions::setSourceLeaseId
      * @covers MicrosoftAzure\Storage\Blob\Models\CopyBlobOptions::getSourceLeaseId
      */
@@ -115,6 +125,8 @@ class CopyBlobOptionsTest extends \PHPUnit_Framework_TestCase
     }
 
     /**
+     * @covers MicrosoftAzure\Storage\Blob\Models\CopyBlobFromURLOptions::setIsIncrementalCopy
+     * @covers MicrosoftAzure\Storage\Blob\Models\CopyBlobFromURLOptions::getIsIncrementalCopy
      * @covers MicrosoftAzure\Storage\Blob\Models\CopyBlobOptions::setIsIncrementalCopy
      * @covers MicrosoftAzure\Storage\Blob\Models\CopyBlobOptions::getIsIncrementalCopy
      */
@@ -125,5 +137,20 @@ class CopyBlobOptionsTest extends \PHPUnit_Framework_TestCase
 
         $options->setIsIncrementalCopy($expected);
         $this->assertEquals($expected, $options->getIsIncrementalCopy());
+    }
+
+    /**
+     * @covers MicrosoftAzure\Storage\Blob\Models\CopyBlobFromURLOptions::setSourceSnapshot
+     * @covers MicrosoftAzure\Storage\Blob\Models\CopyBlobFromURLOptions::getSourceSnapshot
+     * @covers MicrosoftAzure\Storage\Blob\Models\CopyBlobOptions::setSourceSnapshot
+     * @covers MicrosoftAzure\Storage\Blob\Models\CopyBlobOptions::getSourceSnapshot
+     */
+    public function testSetSourceSnapshot()
+    {
+        $expected = '2017-09-19T10:39:36.8401215Z';
+        $options = new CopyBlobOptions();
+
+        $options->setSourceSnapshot($expected);
+        $this->assertEquals($expected, $options->getSourceSnapshot());
     }
 }